+++ CHANGELOG.rst
@@ -17,3 +17,4 @@ Unreleased
 * Add: Simple finite field basic arithmetic
 * Add: Creation of extended finite fields
 * Add: Extended finite field basic arithmetic
+* Add: Reed-Solomon code encoder